Break collapses into separate files.
